use back RunInSeparateProcess
This reverts commit c1b21f580cffe5e60b1376fadae0418d01a540e4.
