--- packages-tests/Caching/Config/FileHashComputer/FileHashComputerEqualsTest.php
@@ -1,39 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\Caching\Config\FileHashComputer;
-
-use Rector\Caching\Config\FileHashComputer;
-use Rector\Core\Configuration\Option;
-use Rector\Core\Configuration\Parameter\SimpleParameterProvider;
-use Rector\Testing\PHPUnit\AbstractLazyTestCase;
-
-final class FileHashComputerEqualsTest extends AbstractLazyTestCase
-{
-    private FileHashComputer $fileHashComputer;
-
-    protected function setUp(): void
-    {
-        $this->fileHashComputer = $this->make(FileHashComputer::class);
-    }
-
-    public function testRectorPhpNotChanged(): void
-    {
-        $this->bootFromConfigFiles([__DIR__ . '/Fixture/rector.php']);
-
-        $hashedFile = $this->fileHashComputer->compute(__DIR__ . '/Fixture/rector.php');
-
-        copy(__DIR__ . '/Fixture/rector.php', __DIR__ . '/Fixture/rector_temp_equal.php');
-        copy(__DIR__ . '/Fixture/rector_rule_equals.php', __DIR__ . '/Fixture/rector.php');
-
-        SimpleParameterProvider::setParameter(Option::REGISTERED_RECTOR_RULES, null);
-
-        $this->bootFromConfigFiles([__DIR__ . '/Fixture/rector.php']);
-
-        $newHashedFile = $this->fileHashComputer->compute(__DIR__ . '/Fixture/rector.php');
-        rename(__DIR__ . '/Fixture/rector_temp_equal.php', __DIR__ . '/Fixture/rector.php');
-
-        $this->assertSame($newHashedFile, $hashedFile);
-    }
-}
